+++ src/semra/io/__init__.py
@@ -1,5 +1,6 @@
 """I/O functions for SeMRA."""
 
+from .graph import from_digraph, to_digraph, to_multidigraph
 from .io import (
     from_bioontologies,
     from_cache_df,
@@ -18,12 +19,15 @@ from .neo4j_io import write_neo4j
 __all__ = [
     "from_bioontologies",
     "from_cache_df",
+    "from_digraph",
     "from_jsonl",
     "from_pickle",
     "from_pyobo",
     "from_sssom",
     "from_sssom_df",
     "get_sssom_df",
+    "to_digraph",
+    "to_multidigraph",
     "write_jsonl",
     "write_neo4j",
     "write_pickle",
